Replace float by double to get rid of artifacts when objects are large.

+++ include/structs.h
@@ -34,9 +34,9 @@ typedef enum e_obj_type {
 
 typedef struct s_vector
 {
-	float	x;
-	float	y;
-	float	z;
+	double	x;
+	double	y;
+	double	z;
 }	t_vector;
 
 typedef struct s_obj
@@ -44,16 +44,16 @@ typedef struct s_obj
 	t_obj_type	type;
 	t_vector	coordinates;
 	t_vector	normalized;
-	float		height;
-	float		diameter;
-	float		radius;
+	double		height;
+	double		diameter;
+	double		radius;
 	t_vector	color;
 	t_vector	emission_color;
 }	t_obj;
 
 typedef struct s_ambient
 {
-	float	ratio;
+	double	ratio;
 	t_color	color;
 }	t_ambient;
 
@@ -62,18 +62,18 @@ typedef struct s_light
 	t_vector	coordinates;
 	t_vector	color;
 	t_vector	emission_color;
-	float		diameter;
-	float		ratio;
+	double		diameter;
+	double		ratio;
 }	t_light;
 
 typedef struct s_camera
 {
 	t_vector	coordinates;
 	t_vector	normalized;
-	float		fov;
-	float		viewporw_size;
-	float		aspect_ratio;
-	float		angle;
+	double		fov;
+	double		viewporw_size;
+	double		aspect_ratio;
+	double		angle;
 }	t_camera;
 
 typedef struct s_ray
@@ -84,8 +84,8 @@ typedef struct s_ray
 
 typedef struct s_hit
 {
-	float		t0;
-	float		t1;
+	double		t0;
+	double		t1;
 	int			temp_part;
 	int			hit_part;
 	t_vector	phit;
@@ -98,22 +98,22 @@ typedef struct s_light_calc
 	t_vector	transmission;
 	t_vector	light_contribution;
 	t_vector	surface_color;
-	float		light_intensity;
+	double		light_intensity;
 }	t_light_calc;
 
 typedef struct s_cyl_inter
 {
 	t_vector	oc;
-	float		dir_dot_axis;
+	double		dir_dot_axis;
 	t_vector	ray_perp_dir;
-	float		ray_perp_len;
-	float		oc_dot_axis;
+	double		ray_perp_len;
+	double		oc_dot_axis;
 	t_vector	oc_proj;
-	float		discriminant;
-	float		sqrt_disc;
-	float		t0;
-	float		t1;
-	float		t_candidate;
+	double		discriminant;
+	double		sqrt_disc;
+	double		t0;
+	double		t1;
+	double		t_candidate;
 	t_vector	intersection;
 	t_vector	to_point;
 	int			*hit_part;
