+++ src/utils/ray_utils.c
@@ -12,7 +12,7 @@
 
 #include "../../include/mini_rt.h"
 
-inline float	length2(t_vector vec)
+inline double	length2(t_vector vec)
 {
 	return (pow(vec.x, 2) + pow(vec.y, 2) + pow(vec.z, 2));
 }
@@ -22,12 +22,12 @@ inline t_vector	revert_vector(t_vector vec)
 	return ((t_vector) {-vec.x, -vec.y, -vec.z});
 }
 
-inline float	dot(t_vector vec1, t_vector vec2)
+inline double	dot(t_vector vec1, t_vector vec2)
 {
 	return (vec1.x * vec2.x + vec1.y * vec2.y + vec1.z * vec2.z);
 }
 
-inline float	lerp(float a, float b, float mix)
+inline double	lerp(double a, double b, double mix)
 {
 	return (b * mix + a * (1 - mix));
 }
